remove 'mr reset' match from mr chart
When the master rating resets, all players MR becomes 0. Then, after the first
ranked match, it changes to 1500 plus variation.

To avoid displaying a nonsensical chart variation like 1876, 0, 1508, I decided
to ignore the 'MR reset' match when displaying mr chart.

Alternatively, instead of ignoring it, I could set the MR to 1500 when it’s
zero, but I can't be sure it will always reset to 1500.