remove the start MR inference after MR reset
Starting from 2024-11-01, the rule for initiating MR at the beginning
of a new phase has changed.
Now, it is based on the final MR from the previous phase.
Since I cannot ensure that the last match in the database is the actual final
match, I have decided to remove the MR calculation from the
"MR calibration" match.