creates a channel to handle async jobs execution that broadcasts result
Creates the cable channel WorkerStreamsChannel to handle cases where
some async processing need to be done, but only after the user connect
to the channel.
It prevent send the response before the user connect to the channel, and
also stop streaming after broadcasting the response preventing some
hanging cable connections.